requirements.txt
pyppeteer>=1.0.2
google-cloud-vision>=3.1.0
pillow>=9.4.0
anthropic>=0.3.0
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-html>=4.0.0
pytest-metadata>=2.0.0
asyncio>=3.4.3
numpy>=1.24.0
requests>=2.28.2
PyYAML>=6.0
colormath>=3.0.0